GenerateEmbeddings/Python/requirements.txt
python-dotenv
pyodbc
azure-identity
sentence-transformers
